Reworked observable mechanism to avoid a memory leak
Observables' reactions used to live in the global store. The problem with this is
that if the observable gets garbage-collected, its reactions would not. They
would hang around in the store forever, leaking memory.

Instead, we store each observable's reactions directly on the object, ensuring
they get freed when it does. This also has the nice side effect of simplifying
a few things, since we no longer have the layer of indirection via Symbols.